llm/pyproject.toml
[project]
name = "llm"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "python-dotenv>=1.2.2",
]
